Move telescope config to telescope plugin file
