+++ lua/plugins/telescope.lua
@@ -17,4 +17,76 @@ return {
       end,
     },
   },
+  config = function()
+    require('telescope').setup {
+      defaults = {
+        mappings = {
+          i = {
+            ['<C-u>'] = false,
+            ['<C-d>'] = false,
+          },
+        },
+      },
+    }
+
+    -- Enable telescope fzf native, if installed
+    pcall(require('telescope').load_extension, 'fzf')
+    -- Telescope live_grep in git root
+    -- Function to find the git root directory based on the current buffer's path
+    local function find_git_root()
+      -- Use the current buffer's path as the starting point for the git search
+      local current_file = vim.api.nvim_buf_get_name(0)
+      local current_dir
+      local cwd = vim.fn.getcwd()
+      -- If the buffer is not associated with a file, return nil
+      if current_file == "" then
+        current_dir = cwd
+      else
+        -- Extract the directory from the current file's path
+        current_dir = vim.fn.fnamemodify(current_file, ":h")
+      end
+
+      -- Find the Git root directory from the current file's path
+      local git_root = vim.fn.systemlist("git -C " .. vim.fn.escape(current_dir, " ") .. " rev-parse --show-toplevel")
+      [1]
+      if vim.v.shell_error ~= 0 then
+        print("Not a git repository. Searching on current working directory")
+        return cwd
+      end
+      return git_root
+    end
+
+    -- Custom live_grep function to search in git root
+    local function live_grep_git_root()
+      local git_root = find_git_root()
+      if git_root then
+        require('telescope.builtin').live_grep({
+          search_dirs = { git_root },
+        })
+      end
+    end
+
+    vim.api.nvim_create_user_command('LiveGrepGitRoot', live_grep_git_root, {})
+
+    -- See `:help telescope.builtin`
+    vim.keymap.set('n', '<leader>?', require('telescope.builtin').oldfiles, { desc = '[?] Find recently opened files' })
+    vim.keymap.set('n', '<leader><space>', require('telescope.builtin').buffers, { desc = '[ ] Find existing buffers' })
+    vim.keymap.set('n', '<leader>/', function()
+      -- You can pass additional configuration to telescope to change theme, layout, etc.
+      require('telescope.builtin').current_buffer_fuzzy_find(require('telescope.themes').get_dropdown {
+        winblend = 10,
+        previewer = false,
+      })
+    end, { desc = '[/] Fuzzily search in current buffer' })
+
+    vim.keymap.set('n', '<leader>ff', require('telescope.builtin').git_files, { desc = '[F]ind git [F]iles' })
+    vim.keymap.set('n', '<C-f>', require('telescope.builtin').find_files, { desc = 'Search all [F]iles' })
+    vim.keymap.set('n', '<leader>fg', require('telescope.builtin').live_grep, { desc = '[F]ind by [G]rep' })
+    vim.keymap.set({ 'n', 'v' }, '<C-g>', require('telescope.builtin').grep_string,
+      { desc = 'Search current word with [G]rep' })
+    vim.keymap.set('n', '<leader>sG', ':LiveGrepGitRoot<cr>', { desc = '[S]earch by [G]rep on Git Root' })
+    vim.keymap.set('n', '<leader>fh', require('telescope.builtin').help_tags, { desc = '[F]ind [H]elp' })
+    vim.keymap.set('n', '<leader>sd', require('telescope.builtin').diagnostics, { desc = '[S]earch [D]iagnostics' })
+    vim.keymap.set('n', '<leader>sr', require('telescope.builtin').resume, { desc = '[S]earch [R]esume' })
+  end
 }
